Add files via upload
I originally intended to make a tool to get data from students and print out which subject needs the most attention but i had pause this project midway as i got to know i had to get a feel about the concepts of 2d arrays. i'll start learning about 2d arrays from now on but in the meantime can anybody finish this project for me?